refactor(home): move followers fetch to server, split tabs into slots
- PeopleFollowersSection becomes async server component fetching followers via getFollowers (Next data cache)
- Split PeopleFollowersTabs into shell + PeopleTab (client) + FollowersTab (server)
- Remove duplicate client-side useEffect fetch and loading skeleton flicker
- Convert (profile-readme)/layout.tsx to server component, extract route-change postMessage logic into RouteChangeNotifier
