+++ apps/home/app/(profile-readme)/RouteChangeNotifier.tsx
@@ -1,0 +1,21 @@
+'use client'
+
+import { postMessageToParent } from '@hamsurang/utils'
+import { usePathname, useSearchParams } from 'next/navigation'
+import { useEffect } from 'react'
+
+export const RouteChangeNotifier = () => {
+  const searchParams = useSearchParams()
+  const pathname = usePathname()
+
+  useEffect(() => {
+    const fullRoute = `/home${pathname}?${searchParams}`
+
+    postMessageToParent({
+      type: 'routeChange',
+      route: fullRoute,
+    })
+  }, [pathname, searchParams])
+
+  return null
+}
