+++ apps/home/app/(profile-readme)/layout.tsx
@@ -1,11 +1,8 @@
-'use client'
-
-import { PeopleFollowersTabs } from '../_shared/components/PeopleFollowersTabs/PeopleFollowersTabs'
 import '@hamsurang/ui/globals.css'
-import { postMessageToParent } from '@hamsurang/utils'
-import { usePathname, useSearchParams } from 'next/navigation'
-import { type PropsWithChildren, useEffect } from 'react'
+import type { PropsWithChildren } from 'react'
+import { PeopleFollowersSection } from '../_shared/components/PeopleFollowersTabs/PeopleFollowersSection'
 import { Playlist } from '../playlist/Playlist'
+import { RouteChangeNotifier } from './RouteChangeNotifier'
 
 export default function PeopleLayout({
   children,
@@ -15,23 +12,12 @@ export default function PeopleLayout({
   readme: React.ReactNode
   profile: React.ReactNode
 }>): JSX.Element {
-  const searchParams = useSearchParams()
-  const pathname = usePathname()
-
-  useEffect(() => {
-    const fullRoute = `/home${pathname}?${searchParams}`
-
-    postMessageToParent({
-      type: 'routeChange',
-      route: fullRoute,
-    })
-  }, [pathname, searchParams])
-
   return (
     <main className="flex gap-6 mobile:flex-col px-4 mt-2 max-w-[1400px] mx-auto">
+      <RouteChangeNotifier />
       <aside className="mobile:w-full w-[296px] shrink-0">
         {profile}
-        <PeopleFollowersTabs />
+        <PeopleFollowersSection />
       </aside>
 
       <section className="flex flex-col gap-4 flex-grow min-w-0">
